SlowDownCode
First Commit has the idea of adding a task to control all
of the
JerkControl, but it appears that it will be just eaiser to use
the 25ms
delay at the end of userControl to judge time. As creating a
time in
embedded C appears to be very complicated... counting cpu
cycles. It is
possible, but requires that the cpu is held at the same
clock speed and
for the clock speed to be known. Which without diving
much deeping than
comfertable into the specs of the cortex brain is
impossible.

:circus_tent: Added .gitattributes & .gitignore files