Jerk Control without new task in usercontrol.
Designed against using seperate task in user mode. There is no point in
adding the complexity of a cpu time for user mode if it is know that it
will be called roughly every 25 ms. So the code was designed instead to
incorperate this and make the motor run up to speed over the course of
two seconds.

In the next commit consideration will be made to override this jerk
control. For times when a burst of energy is needed.