Jerk Control using task's
This branch is going to tackle the problem of jerk control with task. So
it will have cpu times and a more accurate jerk control. However this
accuracy comes at the cost of complexicty. So as of this inital commit
it is unclear weather this style will prevail in the solution of Jerk
Control.