Cleaned Up and Compiled
Before this point the code never had to the chance to be compiled. This
combined with a lack of an IDE lead to the creation of multiple syntax
and function name errors. In this commit all errors and warnings have
been delt with and the code is to be put into testing on a robot.