Sloppy Fix
There was a problem with the previous commit that didn't work. The code would not keep the motors running. This was thought to be the fault a malplaced vexMotorSet(motor(n), 0) within a while loop. However this was not found to be the cause of the problem and something else entirely was causing the issue.